Improve toggling of purchased state
This commit improves the code for toggling of purchased state on bag items.
The improvement is a reduction of complexity by allowing the ShoppingItem
component to fully handle the toggling of the purchased state.
Rather than going through a saga and the global state to handle isLoading
for an individual toggling request, the isLoading state is kept as a
local state within the ShoppingItem. The component is responsible for
triggering the XHR rather than the saga.

This commit also changes how item names are extracted from state for
each bag item. Rather than merging the two lists for a ShoppingList-
specific array, we let them live separately and simply get the names
as needed in the ShoppingList component.