Changes to UI and colors
This commit makes lots of changes to the UI and colors. While the palette
remains the same, green and white colors have been swapped on several
elements. This gives a cleaner and more eye pleasing impression.

This commit also adds proper styling of the input and suggestions for
adding items to the bag.