Pencil SVG, progress Item, improve Input
This commit does a few things, all which are connected.

First, it adds the Pencil SVG. This is to be used by Item in Assortment.

Second, it makes progress to the Item component. This progress is mainly
about using the Pencil SVG and adding the isEditing state. The latter
makes use of the Input component.

Third, the Input component is changed from a stateless function to a
PureComponent, in order to make it possible to set focus in the input
element when the component is mounted.