Feature add item, svg exports and some styling
This commit introduces the feature of adding items to the assortment.
Some styling still remains to be done.

This commit also adds an index file to the svg folder where all svgs are
exported. This makes it a bit easier to import svgs as you get name
completion and you can import several svgs in just one import statement.

Finally, some styling is adjusted with this commit. Nothing big.